Created base test module
